// src/interfaces/common/DomainRole.ts
export interface DomainRole {
    SUPER_ADMINISTRATORS: string;
    ADMINISTRATORS: string;
    USERS: string;
    EXECUTIVES: string;
    MANAGERS: string;
    PROJECT_MANAGER: string;
    ACCOUNT_MANAGERS: string;
    LOGISTIC_MANAGER: string;
    PRODUCERS: string;
    LOGISTICS: string;
    TELESALES: string;
    STAFFS: string;
}